# kivora-payments.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kivora Payments API",
    "version": "2026-07-31",
    "description": "API pública do **Kivora Payments** — cobranças, envios e assinaturas\nprocessados pela Kivora Payments.\n\n## Autenticação\n\nTodas as requisições requerem uma Secret API Key no header:\n\n```\nAuthorization: Bearer sk_test_xxxxxxxxx\n```\n\nou, em produção:\n\n```\nAuthorization: Bearer sk_live_xxxxxxxxx\n```\n\nAs chaves são geridas no dashboard. Nunca exponha Secret Keys em código client-side.\n\n## Versão\n\nA API pública é versionada no path (`/v1/...`).\n\n## Webhooks\n\nA Kivora envia eventos automaticamente para o URL configurado no dashboard.\nNão existem endpoints públicos de gestão de webhooks — apenas entrega de eventos.\nValide a assinatura e responda com HTTP 2xx.",
    "contact": {
      "name": "Kivora Payments Support",
      "url": "https://www.kivorapayments.com"
    }
  },
  "servers": [
    {
      "url": "https://kivorapayments-nu.vercel.app"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "sk_test_... | sk_live_...",
        "description": "Secret API Key. Header: `Authorization: Bearer sk_test_xxx` ou `Authorization: Bearer sk_live_xxx`."
      }
    },
    "schemas": {
...
      "C2BRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "phone",
          "amount"
        ],
        "description": "Pedido para iniciar uma cobrança Customer to Business (C2B).",
        "properties": {
          "phone": {
            "type": "string",
            "description": "Número de telemóvel do pagador (MSISDN local, sem +258).",
            "example": "841234567"
          },
          "amount": {
            "type": "number",
            "minimum": 0,
            "description": "Montante a cobrar.",
            "example": 100
          },
          "currency": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3,
            "default": "MZN",
            "description": "Moeda ISO 4217. Predefinição: MZN.",
            "example": "MZN"
          },
          "reference": {
            "type": "string",
            "description": "Referência externa opcional (ex.: ID da encomenda).",
            "example": "ORDER-1001"
          },
          "description": {
            "type": "string",
            "description": "Descrição legível do pagamento.",
            "example": "Pagamento da encomenda"
          }
        }
      },
      "C2BResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "status",
          "amount",
          "currency"
        ],
        "description": "Pagamento C2B criado ou consultado.",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador do pagamento.",
            "example": "pay_xxxxx"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "paid",
              "failed"
            ],
            "description": "Estado actual do pagamento.",
            "example": "pending"
          },
          "amount": {
            "type": "number",
            "example": 100
          },
          "currency": {
            "type": "string",
            "example": "MZN"
          },
          "reference": {
            "type": "string",
            "description": "Referência externa, se fornecida.",
            "example": "ORDER-1001"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação (presente em consultas GET).",
            "example": "2026-07-30T15:30:00Z"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "description": "Resposta de erro padrão da API pública.",
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Código de erro estável.",
                "example": "PAYMENT_PROCESSING_ERROR"
              },
              "message": {
                "type": "string",
                "description": "Mensagem legível para o cliente.",
                "example": "Unable to process payment."
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/v1/c2b": {
      "post": {
        "operationId": "createC2BPayment",
        "tags": [
          "Payments"
        ],
        "summary": "Iniciar cobrança C2B",
        "description": "Inicia uma cobrança ao cliente (Customer to Business). O pagador recebe um pedido no telemóvel para autorizar o pagamento. O estado inicial é `pending`.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C2BRequest"
              },
              "examples": {
                "default": {
                  "summary": "Cobrança C2B",
                  "value": {
                    "phone": "841234567",
                    "amount": 100,
                    "currency": "MZN",
                    "reference": "ORDER-1001",
                    "description": "Pagamento da encomenda"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cobrança iniciada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C2BResponse"
                },
                "examples": {
                  "default": {
                    "value": {
                      "id": "pay_xxxxx",
                      "status": "pending",
                      "reference": "ORDER-1001",
                      "amount": 100,
                      "currency": "MZN"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Body inválido — phone e amount são obrigatórios",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "default": {
                    "value": {
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Invalid request"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado — Secret API Key em falta, inválida, revogada ou expirada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "default": {
                    "value": {
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Invalid request"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão — a API Key não tem o scope necessário ou o recurso não pertence à organização",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "default": {
                    "value": {
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Invalid request"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Recurso relacionado não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "default": {
                    "value": {
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Invalid request"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "default": {
                    "value": {
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Invalid request"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}